+++ e2e-polybft/cardanofw/test-evm-chain.go
@@ -956,7 +956,6 @@ func (ec *TestEVMChain) DirectBridgingRequest(
 		})
 
 		totalTokenAmount.Add(totalTokenAmount, ra.Amount)
-
 	}
 
 	totalAmount := big.NewInt(0)
